# wal: reduce allocation when encoding entries