@xiang90 2015-03-31
@barakmich @yichengq There are still a lot allocation we can reduce. This is just a starter.